Specify sync interval - force short sleep
The challenge is keeping the future objects in the serial python process and the MPI parallel python process in sync. The user could cancel the future object from the serial python process or the execution of a given future object could be completed resulting in a state change in the parallel python process. For most cases 1/10 of a second seems to be appropriate, still the user can manually adjust this value if necessary.