[major] Refactor the Executor interface (#548)
* [major] Refactor the Executor interface

Replace the backend parameter from the initialization of the Executor class with individual executor classes for the individual backends:
* `Executor(backend="local")` becomes `from executorlib import LocalExecutor§
* `Executor(backend="flux_allocation")` becomes `from executorlib import FluxAllocationExecutor§
* `Executor(backend="flux_submission")` becomes `from executorlib import FluxSubmissionExecutor§
* `Executor(backend="slurm_allocation")` becomes "SlurmAllocationExecutor"
* `Executor(backend="slurm_submission")` becomes `from executorlib import SlurmSubmissionExecutor`

This has two advantages: On the one hand it is less error prone to mistyping the backend name, as the user can use auto completion to import the right module. On the other hand it is more consistent with the standard library which defines the `ProcessPoolExecutor` and the `ThreadPoolExecutor`, rather than a `backend` parameter.

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix test

* update benchmark

* move interfaces to individual modules

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fixes

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* remove create

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* remove redundant functionality

* Move graph tests

* remove core check

* clean up

* more tests

* fix import

* more fixes

* Rename LocalExecutor to SingleNodeExecutor

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* test block allocation

* fix

* flux fix

* Rename interfaces

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* fix docstrings

* add docstrings to create_*_executor() function

* fix test

* merge

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>